Shifted AnswersList's state to Question
